Omit vanilla tests for keyset
Summary:
We are not currently targeting keysets for bespoke specialization. To reduce the number of vanilla checks necessary, we never emit logging arrays for keysets. Because keysets are never specialized, all keysets are then bespoke and we can assert that any value known to be a keyset is also vanilla.

We also avoid allocating profiles and emitting NewLoggingArray for keysets.

The logic to determine what DataTypes can be bespoke is contained in `arrayTypeMaybeBespoke` for extensibility.

Differential Revision: D25264755

fbshipit-source-id: f5ad5d18d6639ce4df7d87d3fcf0f5a7e2ef28eb