inline Decl_ancestors.make
Summary:
I wish to get tighter rein over linearizations. I want to know precisely when a linearization is being computed. This is towards my end-goal of moving linearizations into the decl service.

This diff is a small step along the way. It moves creation+use of linearizations out of Decl_ancestors into the sole consumer, Typing_classes_heap. It eliminates a temporary structure Decl_ancestors.ancestor_caches which was used to communicate between the two, and is now not needed.

Differential Revision: D24669119

fbshipit-source-id: 975c4c1a7c9ecd83050e11c600f9508d86a82b13