refactor Decl_linearize.linearize
Summary:
This is a trivial refactoring.

Prior to this diff, `Decl_linearize.linearize` did work to assemble the first element, and also to assemble the list of ancestors. I thought it looked cleaner when that work is done inside the appropriate part of the state machine. It means that the `Child` state no longer needs to carry around a list of ancestors.

Differential Revision: D25180952

fbshipit-source-id: 62f8be356be9cbc89425f35713eab16715e1e9bc