Decl_linearize.get_linearizations (both at once)
Summary:
I will want to encapsulate the two linearizations as a pair.

That's to allow (1) if the implementation is easier when it produces both linearizations simultaneously, (2) if I want the construction of say lin_ancestors to be contingent upon the results of lin_members. (Jake suggested that we could do Stringish more efficiently this way).

Differential Revision: D24669190

fbshipit-source-id: de65bd3a39154ed70f9305bbf3b85fc6d0a6838e