Initial iter-specialization support for bespokes
Summary:
We support specialization for monotype vecs and monotype dicts.

To do so, we add a two-byte "layout" field to the Iter struct. This field is NOT the concrete layout of the array being iterated over; it is the (possibly abstract) layout that we specialized code for. We set this field when specializing IterInit for bespoke array-likes and check it at IterNext. In most cases, we can eliminate this check in load-elim.

Other than that, we simply add a BespokeAccessor type to dispatch to ArrayLayout methods. The code generated this way is...well, it's hot garbage, but we'll fix that by improving type inference and adding new monotype specializations in a follow-up.

Differential Revision: D25279829

fbshipit-source-id: 73d09052dcff48c25f1f3b23c27fdd6c9ce5c115