Remove .hhconfig experimental feature for IFC in favor of a file attribute/unstable feature
Summary:
This diff converts the .hhconfig option for IFC into an unstable feature instead. This allows us to fail at the parser level when someone tries to use IFC, instead of only at the typechecker level. It also allows us to turn on IFC for a file at a time, instead of only on policied functions.

This diff depends on us removing the .hhconfig option before landing, since otherwise master will not work on www.

Reviewed By: zilberstein

Differential Revision: D25032154

fbshipit-source-id: 9ad9172af8af3cb670cee3c82f615ddc0a92e1d5