Delete virtual ArrayLayout irgen methods
Summary:
We've decided to generate good bespoke code by emitting generic ops and repeatedly simplifying based on input type information. This approach has the advantage that we'll generate good code regardless of when we get layout information during analysis; it's a pre-requisite for specializing inlined versions of highly-generic functions.

As a result, we can delete every one of the ArrayLayout and BespokeLayout "emit..." functions. The default implementations of most of these functions was a single HHIR op, anyway. The main question is whether we still get the specialized MonotypeVec code. I claim that we still produce the right op due to our simplifiers, and that when mcolavita adds (virtual) type inference functions, we'll have even better code than before.

Differential Revision: D25319355

fbshipit-source-id: dfed6a839152e156323013f9f7567f3fe43f88a1