Iterable/Traversable support in foreach
Summary:
In addition to Hack array types, anything that implements Traversable/Iterable can be traversed using `foreach`. This diff adds the support for that.

Because we distinguish key and value policy types in Hack arrays but conflate all fields in classes, traversing over a Traversable is less precise than an Hack array type.

Reviewed By: mpu

Differential Revision: D25056554

fbshipit-source-id: eccb838d9a5971be9d358e857029e61774fec369