Refactor array like types logic
Summary:
We had a generic way of gathering just Hack arrays or Hack arrays and shapes together. This will need expanding soon with tuples and classes that implements `KeyedContainer`. In anticipation of this, this diff refactors the logic to accommadate more than two variants in an extensible way.

The main benefit of this centralised code is to handle intersection uniformly and to handle defaults all in one place.

Because flags now determine which variants will be in the result, we have some catch all pattern matching behaviour that should never occur if the options are set right.

Reviewed By: mpu

Differential Revision: D25186588

fbshipit-source-id: 6696a51f6732fc07b3ed1f2d7cfd2b9a8b4cbb50