automatically log data to scuba after sampling
Summary: Before this diff, when we want to record the memory of running `f`, we wrap it with a function `collect_cgroup_stats` to record the data, then we call `log_to_scuba` to log the data to Scuba. This diff combines the two so that we only need to call one function at the call site.

Differential Revision: D25220710

fbshipit-source-id: b955814649a791bf8071b54e82de617314ffaa5d