Optimize RepoOptions::cacheKeySha1()
Summary:
Everytime we call RepoOptions::cacheKeySha1(), we mangle all
of the options together into a string, then hash it. However, since
the options inside a given RepoOptions cannot ever change, we can just
calculate the hash once (when its created), and just store that and
return that. Moreover, we can speed up the calculation a bit by
appending to the same string, instead of creating sub-strings and
concatenating everything together (better still would be incrementally
calculating the SHA1, but we don't have an existing API for that).

Reviewed By: oulgen

Differential Revision: D25326664

fbshipit-source-id: 4d5a199bddaa162f63b55ea14158362f7f69daa1