Always assert TCA is not nullptr for smashable instrs
Summary:
Debugging a core with bad prof data, kshaunak and I found
```
   0x000000000de42c9f:  callq  0x60280a <HPHP::MixedArray::ExistsStr(HPHP::ArrayData const*, HPHP::StringData const*)>
   0x000000000de42ca4:  test   %al,%al
   0x000000000de42ca6:  je     0x0
   0x000000000de42cac:  incl   0x0(%r13)
```
This shows that we make a MixedArray::ExistsStr call, it returns false, and then we jump to 0. Notice that the jump is 6 bits wide which means that it used to be a big jump that possibly got smashed.

Lets make sure we always assert to prevent such things.

Reviewed By: binliu19

Differential Revision: D25346717

fbshipit-source-id: b7f8ccc9adb847ccf88ce8f17f37b6b5c879e470