Use a vanilla layout at unsampled sinks
Summary:
All "unsampled" arrays are coming from sources that we don't know how to convert to bespoke sources. As such, we should choose vanilla for them.

If a given sink only sees unsampled arrays, we'll already make it a vanilla sink. However, if it seems a very small number of sampled arrays, we'll currently make it a top sink. Fix that logic.

Reviewed By: oulgen

Differential Revision: D25362204

fbshipit-source-id: 4dc41c8a870be15187552ec259107cd91653f270