Change signature of Errors.add_list
Summary:
Errors.error has a `claim` part and a `reasons` part. In the future we'd like the claim to be a `pos_from_ast message` while reasons would be `pos_from_anywhere message`s, which means we can't carry around uniform lists of messages anymore.
This diff is a first step in this direction. It changes the signature of Errors.add_list so that it does not accept a uniform list of messages anymore, but instead a claim and a list of reasons.

Reviewed By: Wilfred

Differential Revision: D25333935

fbshipit-source-id: 3fe3e5b4068ff24fff0ccf1ab98c4a58f34a2ebd