Fix crash when passing empty auto_namespace_map to direct decl parser
Summary:
UnsafeOcamlPtr is for pointers to block values. Since it represents a data structure whose lifetime not statically known, there's no sense in using it for immediate values (an immediate value could just be read as an integer instead).

This is a problem in rust_decl_ffi, where we use UnsafeOcamlPtr for the namespace map (which is a list). A non-empty list is a block value, but an empty list is an immediate value, so we crash here when the direct decl parser is invoked with an empty auto_namespace_map. UnsafeOcamlPtr::from_ocaml returns an Error, we unwrap it, and `ocaml_ffi` converts the panic into an OCaml exception.

The use of so many `UnsafeOcamlPtr`s with `ocaml_ffi` is a bit of a code smell. `ocaml_ffi` is a macro which adds FromOcamlRep conversions for the arguments, adds a ToOcamlRep conversion for the return value, and catches exceptions in the wrapped function using `ocamlrep_ocamlpool::catch_unwind`. Since UnsafeOcamlPtr has trivial implementations of FromOcamlRep and ToOcamlRep, it can be used like an escape hatch from this conversion behavior in `ocaml_ffi` when used for one of the arguments. When we use it for *all* of the arguments and the return value, however, we've opted out of `ocaml_ffi`s behavior almost entirely, and should probably just use `catch_unwind` directly instead.

Differential Revision: D25288403

fbshipit-source-id: c73e99dcaa6b9f8fe77dbb76c694530b8f0735e4