Update sub-typing for sound dynamic
Summary: Only allow a type to be a sub-type of dynamic where it is safe to do so. This includes a variety of built-in types, and classes that explicitly implement dynamic. Updates to the typing of expressions and the checks that a class can safely implement dynamic are to come. For now don't let dynamic be a subtype of anything except for mixed. We will relaxing this where HHVM enforces a type hint, eg, passing a dynamic to a function with an int parameter. All changes protected by a --enable-sound-dynamic-type flag.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D24923122

fbshipit-source-id: d6d42d0d38d6b233e14c4f7cf590900f51ff7607