Fix offsets in tokens returned by with_leading and with_trailing
Summary:
In {D24473983 (https://github.com/facebook/hhvm/commit/55780b3de39f4d3058b17fb37c0f5a7ef23af6a1)}, Steve fixed a bug in PositionedToken::with_leading--we were failing to update the `offset` and `leading_width` fields to account for our new leading trivia. Since the `offset` field is meant to point to the beginning of the leading trivia, and the `leading_width` is meant to be the width of the leading trivia in bytes, we're in an inconsistent state if we replace our leading trivia but fail to update these fields.

This diff applies the same fix to CompactToken and the owned PositionedToken type (Steve's diff only fixed the bug in the by-ref PositionedToken). It also updates `trailing_width` in `with_trailing` for both PositionedToken types.

Reviewed By: shiqicao

Differential Revision: D25383084

fbshipit-source-id: c2d3f62572afe01e059282f05ccc35b14a4cb4f5