Correctly infer type of parenthesized strings
Summary:
A missing case here caused us to fail to infer the type of string expressions when parenthesized.

We can also delete the position-rewriting in make_parenthesized_expression--it helped match the behavior of OCaml decl when class consts contained their expressions, but they no longer do. It's also the wrong behavior for the added test case, where the position points to the string literal and does not contain the surrounding parentheses.

Reviewed By: shiqicao

Differential Revision: D25383083

fbshipit-source-id: 62272bd9b2c0675de29d89676652deb08657d092