Avoid putting bespoke arrays in APC
Summary:
A key requirement on our sampling strategy is that we attempt to sample each array constructor with a Bernoulli distribution. Unfortunately, arrays that are stored in APC and then retrieved violate this simple constraint - samples at a given site become highly correlated based on whether the particular APC array stored for that location was a LoggingArray or not.

I have a long-term fix for this issue, but we won't be able to implement it quickly. In the meantime, these APC arrays are causing two key problems:
- For locations where we get no samples at all, we'll do Top codegen.
- For locations where we get one or two bespoke array samples, we'll do bespoke codegen and then side exit from the optimized translations.

Let's convert arrays to vanilla, unsampled arrays when we store them in APC, and count any sinks downstream as vanilla sinks.

Differential Revision: D25364097

fbshipit-source-id: cc9cfb6254903bb6800f42e0cddb5b90ca3cd1a9