Add helper to recursively tag collections
Summary:
Per request by viratyosin, this helper would help us with certain cases where we mix array-likes and collections. The trick with collections, of course, is that we have to deal with cyclic cases. Here's the basic idea:
1. We keep a hash table of array-likes and collections that we visit.
1. The first time we visit an array, before mutating it, we store a cached result of nullptr for it.
1. After mutating it, we cache the real result.
1. If we ever visit an array and we've got a nullptr cached for it, we copy it, immediately cache that copy, and mutate the copy in place.

Crucially, we rely on the fact that we can mutate a refcount-1 array in place to allow us to return the copy as a cached result before we've finished mutating it.

As a runtime bound, note that we'll only ever mutate an array from the original input twice, so even if we have some complex recursive data structure, the overall runtime is linear in its total size.

Reviewed By: dneiter

Differential Revision: D25380346

fbshipit-source-id: 6069066837fd6a5926654267fb452d7848389f27