Address serialization of exception traces
Summary:
Every PHP Error or Exception object has a private property called $trace which is composed of PHP arrays: it's a varray of stack frame darrays. We don't want to mark all of these backtraces on construction - that would be a code hygiene disaster! Instead, we resolved this problem for the most part by tagging the property on access (via the getTrace() call).

However, there are a few ways that private properties like $trace can leak WITHOUT being exposed through a safe helper. We could serialize the exception object (ugh), or we could cast the exception object to an array (double ugh). It turns out that serialization is a special case of the array cast, since that's how it's implemented internally.

To deal with marking in these cases where no Hack code explicitly accesses $trace, we add logic to the array cast helper (exposed to Hack as object_prop_array) to mark the $trace field. We also need to install marking logic in one other place: if the $trace is a compact backtrace, it's only materialized into an array at the point it's serialized, so we mark it there.

Reviewed By: ricklavoie

Differential Revision: D25417423

fbshipit-source-id: 656987c5193f8c4f14846e7b007d1786789b645f