Xml node rewriting: Can't eliminate a clone() but can improve readability
Summary: There was a hope that we could refactor `rewrite_xml_` to eliminate some clones. That turns out not to be possible. Nonetheless, it was possible to improve readability in that area so putting that work up for landing. As expected, it hasn't affected performance of emitting either positively or negatively.

Reviewed By: shiqicao

Differential Revision: D25466189

fbshipit-source-id: 234f9a04594c395de7460eae6060a307311fde6f