Update serde_json version in Cargo.lock and regenerate autocargo files
Summary:
The fbcode Cargo.toml has recently changed to include the float_roundtrip feature for serde_json. However, our Cargo.lock is currently pinned to an older version of serde_json which doesn't provide the feature. As a result, regenerating our Cargo files results in build errors--Cargo complains that the version of serde_json we're pinned to does not provide the requested feature.

We can just upgrade to a newer version of serde_json--then we should be able to use the regenerated Cargo files which ask for the float_roundtrip feature.

Reviewed By: shiqicao

Differential Revision: D25481313

fbshipit-source-id: c3d065cc66c6fafe205b473e2c8c9f807fc3c8aa