Add telemetry to recovery from breaching heap cap
Summary: We already record the size of the OCaml heap at the end of processing a bucket of files, in `final_heap_mb`.  We now add `max_heap_mb` for the maximum size that the heap reaches before any attempt to flush caches and do a major GC, and `exceeded_cap_count` for the number of times during the processing of each bucket that the memory cap was exceeded.

Reviewed By: ljw1004

Differential Revision: D25494228

fbshipit-source-id: 7a904e2d656687398d4ed712fb52a6af14f67be6