Attempted cleanup of translation creation logic
Summary:
We have a ton of paths for creating translations spread over many
files.  Notably for both tracelets and prolouges we have 3 paths:
- Live/profile generation
- Optimized retranslation
- Optimized retranslation through retranslate all

Add to this the logic to handle relocation based on the (possibly) broken
ReusableTC option and the relocation for parallel local code emitting (used
during retranslate all).  All in all we end up with a total mess.  This is an
attempt diff to move all translation to some common infrastructure based on a
phased translation approach, where the majority of the logic is common between
prologue, tracelets and any future translation type.  Furthermore we drop the
specialized relcoation logic for ReusableTC, and use the same setup used for
thread local code emission.

Reviewed By: ottoni

Differential Revision: D25281945

fbshipit-source-id: 0c52bff95e3cfc5733d39aa992eeba0c44d55e49