Use mmap to reserve the arrprov slab
Summary:
We allocate provenance tags by reserving a single 16Gb slab (8Gb in lowptr builds) and then bump-allocating tags from it.

Our goal here is to keep the resident set size small. We should only pay for the prefix of this slab that we actually use. Using malloc to reserve the slab happens to work in prod, but that's only because optimized allocators don't initialize the allocations they hand out. However, it breaks when we use a debug-mode allocator, as in D25157799 (https://github.com/facebook/hhvm/commit/b936630b2dc73d7f1386c9117bdd7dcc59b14e84), because the debug-mode integrity checks cause us to access all of the slab's memory.

We should not rely on malloc's behavior here. As ricklavoie notes, since we're not going to free this memory, there's no reason to use an allocator at all - we can use mmap.

Reviewed By: alexeyt

Differential Revision: D25515629

fbshipit-source-id: 24d027977c0636e169e0f2f83e1a4d8fba78f2f1