Collect argument inspection to a common function in special builtins
Summary:
This diff stops special builtins from inspecting the type using topT and popC and collects these inspections under the getArg function. The goal of this is that when we stop pushing the default arguments, this getArg function gen_const them.

The assertions will be done inside this getArg function.

Reviewed By: ricklavoie

Differential Revision: D25520847

fbshipit-source-id: b2714a4c11a060db78dcb618e242fc5ddf2518ed