Dont emit FCallBuiltin unless HHBBC optimizes the function
Summary:
In order to unify the calling convention coeffect enforcement, I'm killing FCallBuiltin. This diff stops HHBBC from converting FCallFuncD to FCallBuiltin unless HHBBC is doing additional optimizations.

Same folding for builtins are also done in FCallFuncD's interp step and I'm planning to convert the remaining FCallBuiltins by moving the rest of the optimizations to FCallFuncD.

As for the JIT, builtins get unconditionally inlined with no cost analysis and as a result of this, FCallBuiltin doesn't give us an perf wins, we end up emitting identical code.

Reviewed By: ricklavoie, jano

Differential Revision: D25450845

fbshipit-source-id: be4b206b5c163447f6dbd7f8d733cb488d7c141c