Do not allow inout, unpack or variadic arguments in FCallBuiltin conversion
Summary:
Currently, we are only converting the special native functions to FCallBuiltin. None of these use inout, unpack or variadic arguments. Lets kill the support for it to further simplify the code.

On the next diffs, I'll update the logic so that we do not push the default arguments for the sake of inspection but operate over types.

With the goal of killing off FCallBuiltin.

Reviewed By: ricklavoie

Differential Revision: D25519515

fbshipit-source-id: da5e95d932b5b34536f0205ab61656fa99be6592