Move stack modifications to handle_builtin
Summary:
This diff moves all modification to stack values to handle_builtin. Each of the `builtin_` functions now return TypeOrReduced which means that if the builtin returns Type, then we push that to the stack, if it reduces, we do nothing and if it didn't reduce and didn't return a type, then it must mean there wasn't an optimization.

This logic will make it so that FCallBuiltin to FCallFuncD diff will not have to modify the optimizations for these special builtins.

At first I tried using Either<Type, bool> but that only works for pointer types.

Reviewed By: ricklavoie

Differential Revision: D25520958

fbshipit-source-id: 7784253db5657c2240cb9297885983640782358b