Typing_check_service uses BigList
Summary:
Typing_check_service manipulates "files_remaining", a list of all files remaining to typecheck. In a full init, this will start with every single file in the repo, which might be huge. Therefore this variable is costly to traverse.

This diff is a no-op. It simply uses BigList.t instead of List.t, as a strongly-typed way to indicate in the code which parts a costly.

Reviewed By: yinghuitan, madgen

Differential Revision: D25523300

fbshipit-source-id: 2ebb7c046a915c41a21cc918d652e3834d6071f1