naive "implements dynamic" soundness checks
Summary:
This diff sets the basis for correct but naive soundness checks for when a class declares that it "implements dynamic":

- if a class implements dynamic, then all its subclasses must implement dynamic as well;
- if a class does not implement dynamic, then none of its superclasses can implement dynamic;

- if a class implements dynamic, then all its methods are dynamically callable.

For this diff, a method is dynamically callable if all its arguments are enforceable, and its return type is coercible to dynamic.

Future diffs will complete this diff, addressing at least the following limitations:
- interfaces and traits are not yet taken into account when performing the inheritance checks;
- the handling of functions returning void is too strict;
- reified generics are ignored for the enforceability check.

This diff should have no impact on the hack typechecker unless the option --enable-sound-dynamic-type is selected.

Reviewed By: vsiles

Differential Revision: D25397164

fbshipit-source-id: effe33f4ff192b45248e652e8311d22d4764b7bf