use lists not Sequence.t for linearization generation
Summary:
I have a hunch that the benefit of lazy linearizations (only do the minimum work necessary) is outweighed by the disadvantages (overhead of Sequence.t, and the disadvantage that every worker has to duplicate work until the first one manages to complete the sequence).

I wrote an earlier experiment D25173283 which used the existing Sequence.t behavior but changed it so that, as soon as someone asked for even the first element of a linearization, it just went ahead and got all of them and wrote the complete linearization to shmem. Perf results there showed that the experiment didn't harm perf, and might have yielded a slight perf improvement. Conclusion: lazy sequences aren't showing benefit.

And I wrote a yet earlier diff D24622099 (https://github.com/facebook/hhvm/commit/5fa6f816e421f59c53031c87d2270fb86078644a) which wrapped up the entirety of sequence creation (not sequence elaboration; merely sequence creation) behind a lazy; that diff showed a 1-2% benefit. Conclusion: although our big-O notation cost for sequences looks good, they might hurt just in the constant factor.

NOW FOR THIS DIFF

Given that laziness of linearization doesn't help, and Sequence.t appears to have overhead, let's try eliminating it. This diff switches to linearization-creation to list rather than Sequence.t.

It still exposes it publicly as a Sequence.t. That's so that members can be lazy. My hunch is that the cost tradeoff for members is worth it, and they benefit from laziness.

It still retains the "move_next" state machine for generating the linearization. I think this is a very nice piece of code, and I'd like to port the same code structure to Rust, so that the decl-service can one day do linearizations lazily in a way that doesn't have the downsides of ocaml.

The perf results suggest this hunch has paid off, with a 2%-4% perf win.

Differential Revision: D25315046

fbshipit-source-id: 236e75172f4e0b3d20e0cf73d67a52eef07ada25