Array inout builtins preserve markings
Summary:
Certain builtins take array-likes by inout and are logically mutations on the inputs. Here's a list:
- all the sorts, including multisort
- array_push/pop/shift/unshift
- array_shuffle
- array_splice

All of these inputs should propagate the legacy bit: the output should be marked iff the input is. Today, if the input is unmarked, so is the output; but if the input is marked, the output MAY or MAY NOT be marked. What a mess! We fix them all, and add a comprehensive unit test. We also plug a related array provenance hole: when we unmark a dvarray, we reassign its provenance tag.

Reviewed By: arnabde03

Differential Revision: D25550783

fbshipit-source-id: 754a41a1270321296a35d1b1b02feeed7bcedab2