Allow a user-defined adapter to convert between runtime type and thrift type
Summary:
To support hack.adapter annotation (roughly equivalent to cpp.type), if 'adapter' field is specified in the field spec, call `toThrift`/`fromThrift` methods during serialization/deserialization correspondingly. This provides a way to decouple how the data is represented in memory vs on the wire (eg a timestamp might be represented as `DateTime` at runtime, but int64 timestamp on the wire)

Works for everywhere (structs, primitive types, containers, etc) except for map keys when hack arrays are used (b/c they can only be int/string and there are some extra checks for types there that would be painful to work around). Probably going to disallow the adapter annotation on map keys in general.

Reviewed By: Alfus, andrii-korotkov

Differential Revision: D25139059

fbshipit-source-id: 35c29a8f5128919034b3ecaaf87ab08fbac3f3ce