Calculate type-constraint effects for static property mutations
Summary:
Take into account the possible effects arising from type-constraints
on static properties. With this information we can determine if a
given static property mutation can possibly throw or be considered
effect-free. In addition, it lets us push more accurate values onto
the stack as a result of the set.

Reviewed By: jano

Differential Revision: D25484305

fbshipit-source-id: 3a575820d7c9a4ed49b2b5dfc0ad88407e185459