Refactor static property interfaces in hhbbc
Summary:
Currently static property information is stored in a number of places,
and there's no real centralized logic to retrieve it. There's a number
of ad-hoc functions for every individual piece of information you want
to know.

Refactor these interfaces. Now there's just two functions on the
 Index: lookup_static() and merge_static_type(). lookup_static() is
responsible for taking types representing the class and property name,
and determining which property it might represent and returning *all*
relevant data at once. Unlike the previous logic, this deals with
ambiguous cases where there could potentially be multiple matching
properties. The caller does not need to worry if the property is
public or private, the function takes care of this itself, applying
the necessary visibility rules. For mutations, merge_static_type() is
similar, except it handles merging in a new type to all possible
matching properties.

lookup_static() returns a struct containing relevant information about
the property access, including its best known type and whether it
could potentially throw (broken down to different reasons). With this
information we can more reliably determine when a static property
access is effect_free, or can be const folded.

Modify the interp functions to use the new interfaces and leverage
this new information. This allowed for the cleanup of a lot of dead or
now unneeded code.

Reviewed By: oulgen

Differential Revision: D25459911

fbshipit-source-id: 4e7de42621c5911cd6f99deaf19bf42361e129cb