Improve block merging in HHBBC
Summary:
While looking at HHBBC output, I realized that some instruction
sequences weren't getting strength reduced like they should have. It
turns out that this is because the sequence spanned a block (we only
strength reduce within a block). The only reason they spanned a block
was because of a change in the destination of the exception edge. In
many of these cases, however, the sequence itself was all
non-throwing.

Improve the block merging logic. We cannot usually merge two blocks if
they have different exception edges. However, if at least one of the
blocks is entirely no-throw (we already know this from analysis), we
can merge them anyways. If the block can't throw, its irrelevant what
its exception edge is. By merging the blocks, it gives more
opportunity for the strengh reduction rules to fire.

One catch with this is that occasionally it can produce bytecode which
the verifier will reject. This is because the verifier does not
realize the sequence is no-throw (this requires full analysis
generally) and thus will think the exception edge can be
traversed. Merging blocks might cause iterator state violations or
silence level violations. If the blocks to merge contain any of these
relevant instructions, do not merge them.

One additional optimization: If a block's exception edge is just to
another block which immediately throws again, we can rewrite the
exception edge to the second throw's target. We repeat this process
until we either hit a non-trivial catch block, or leave the unit. This
is essentially jump threading, but for exception edges. Not only do
this reduce the number of redundant catch blocks, it makes it more
likely for adjacent blocks to have the same exception edge.

Reviewed By: oulgen

Differential Revision: D25511735

fbshipit-source-id: 036b1b29cdbf0197ebc05cfc0d5f18ecd26e3d75