Add TVecish and TDictish types to HHBBC
Summary:
TVecish is the union of TVArr and TVec. TDictish is the union of TDArr
and TDict. These unions are useful because vec and varray and dict and
darray behavior very similar to one another in most circumstances.

Modify the array helper functions to now break them down according to
vecish or dictish (and keyset). The array-like one still remains, to
deal with unions. There's no longer any need to have a plain array
one, since they'll be treated as a union of vecish and dictish by the
array-like function.

Export these as RepoAuth assert types as well. I noticed that we
weren't exporting ArrLike either, which gets inferred quite a bit, so
do that as well.

Differential Revision: D25511170

fbshipit-source-id: b64fc8ba40302e978ef7a8c082e15855db9a8335