Some renaming
Summary:
- {f,m,hf}_cap -> _ctxs
- {f,m}_unsafe_cap -> _unsafe_ctxs
- Syntax.Capability -> Syntax.Contexts
- various local vars and functions
- updates schema version because we missed that in the last few diffs

The cap naming is leftover from the provisional syntax where the cap was just a type, and we'd use the intersection syntax to do multiple.  Now the distinction is clear: the concrete and abstract syntax trees store a list of contexts, which are simple type hints, and function type records (decl or locl) store a single capability type.

Reviewed By: DavidSnider, losvald

Differential Revision: D25551684

fbshipit-source-id: a4bf8ffd2c04dd04fc6daf1108af4880951a5875