Introduce make_tv_of_type helper
Summary:
colavitam identified a GCC bug that hurts the performance of our code in a big way. Basically, any unitialized struct fields can cause a ludicrous amount of spilling.

This diff simply replaces all of the offending constructor calls with a wrapper, so that when we introduce a workaround, we only have to fix one site.

Reviewed By: colavitam

Differential Revision: D25608210

fbshipit-source-id: b6042f975c7ad4c2c4ed9815d6835ce5f080f6e6