Lower `ctx $f`
Summary:
Given a
```
function f(
  (function (tbar)[_]: t) $f
)[ctx $f]: void {}
```
the parser does a transformation to
```
function f<Tctx$f> (
  (function (tbar)[ctx $f]: t) $f
)[ctx $f]: void {}
```
The `Tctx$f` is a new, non-denotable type parameter. `ctx $f` remains a hint in the AST until it is used, at which point it is converted to a decl ty `Tgeneric "Tctx$f"`. In other words, the typechecker handles `ctx $f` hints **under the assumption** that the parser has inserted a type parameter named `ctx$f` in the function's type parameter list.

We could replace the `ctx $f` hints with generic type hints, but this way makes compilation simpler.

Reviewed By: DavidSnider, losvald

Differential Revision: D25573746

fbshipit-source-id: 8c2f02f8344f08b3ea4f6f32cf264c376541e1bd