Add unit test for typehint reflection
Summary:
We've been saying all along that typehints accessed via type_structure or ReflectionClass won't change with The Big Switch, but I realized that I don't know of a test for that. Here it is.

We don't need to explicitly enable HADVAs for it - we have a separate HADVAs contbuild.

Reviewed By: arnabde03

Differential Revision: D25638669

fbshipit-source-id: f2d9ed09f999ba099ae8a4c2dc52deffd97f5e18