Declare folded classes in Decl.make_env when using DD parser
Summary: When the direct decl parser is enabled, and shallow decl is disabled, Decl.make_env should produce folded classes and write them to shared memory (it is used specifically for this side effect). We overlooked this, though--at the moment, it only writes shallow classes to shared memory. This diff iterates over the list of shallow classes produced by decl-parsing the file and produces a folded decl for each.

Differential Revision: D25597994

fbshipit-source-id: 44997d1bd9686210c4d7e10d3b916494a040b594