Fix HADVAs bug: reified generics typechecks
Summary:
Oof. If we didn't fix this bug, every time we enforce a varray or darray typehint, we'd fatal post-HADVAs.

Note that, like varray and darray typehints pre-HADVAs, we allow both vecs and dicts to past post-HADVAs. That behavior isn't ideal but it's something we have to address post-unification when we fix typehints. It shouldn't be too bad to address it at that step - these typehints are mostly obeyed.

Reviewed By: arnabde03

Differential Revision: D25640318

fbshipit-source-id: 4f0e8bc1cd4813d8557cafae26f32ea6f3a7de73